rename: type.ts 위치 변경
