+++ src/ai/api/studyPlan.ts
@@ -1,5 +1,5 @@
 import { http } from './http';
-import type { StudyPlanListRes } from './types';
+import type { StudyPlanListRes } from '../types/types';
 
 export function getStudyPlans(userId: number, limit = 100, offset = 0) {
   const qs = new URLSearchParams({ limit: String(limit), offset: String(offset) }).toString();
